Parallelized set import for performance
Still a WIP, but parallelizing the data import (one goroutine per set)
reduced the entire dataset import time from 3.5 hours to 30 minutes.

Needed to tune the idle connection limit for the db to prevent thrashing
connections (and running out of local ports because of all of the dead
tcp connections in CLOSE_WAIT), and needed to introduce some locking to
prevent interleaving db writes in the middle of partial reads (mysql
limitation).  Still some more optimization to come