Large performance improvment to db import, other fixes
* Removed janky (although clever, I thought) use of RWMutex to protect
against interleaving of inserts and selects on same db connection.
Instead, moved to using separate db connections per-goroutine, which is
probably the correct solution
* Wrapped all db operations per-set goroutine in transactions, which
massively improved import performance.  Suspicion is that previously, db
was naively commiting every inserted row to disk as it came in, leading
to lots of blocking on disk latency.  Using a transaction causes large
batches of inserts to be commited to disk in a block, improving
performance considerably.  The db could probably be tuned to do this
without the use of transactions, but since I was planning on adding
transactions anyway at some point, this kills two birds with one stone.
* Additional small modifications to the db schema, and continued
refactoring of import logic to make it more encapsulated