Worked on proof of concept for an alternative import scheme
I was having lots of issues with concurrency in the previous import
scheme, specifically around properly locking access to the atomic card
data table.  Because multiple cards can reference the same atomic card
data, making the import for each set concurrent was running into issues
where multiple versions of the same atomic card data record were being
inserted into the db, leading to inconsistent db results between various
runs of the importer.

To try and fix these issues, I'm trying an alternative scheme, where all
accesses to the atomic_card_data table are serialized through one
goroutine, and all of the other goroutines make requests for atomic card
data to this one routine.  This eliminates the need to use explicit
locking mechanisms on the table, either in sql or in the application.

This appears to work, a few runs of the importer leave me with a
consistent database, but unfortunately is barely more performant than
just importing all of the sets sequentially and not using any
concurrency, which leads me to believe that the importer routines are
spending most of their time blocked on the atomic_card_data goroutine.
Committing this for now, and I'll keep working on performance.