Refactored card db schema and associated import code
After fighting with it for a bit, I came to the realization that
splitting the atomic card data into its own table just wasn't worth it.
For the cost of saving about 4k rows in a table with ~1.7M rows, I was
giving myself all sorts of locking and concurrency headaches, not to
mention more complicated queries for card data from consumers.  This
refactor puts all of main card data into a single table, allowing much
easier concurrent imports, since each card record is now unique, so
access to the db doesn't have to be synchronized.