Turned card-importer into periodic service
card-importer is now a periodic service, that when started, will check
for updates every 12 hours.  It also responds to SIGUSR1 to request an
immediate update check, and SIGINT, to cleanly shut down if it's in the
middle of an update.  The plan is to packate this into a docker
container, where it will run as a microservice that keeps the card and
price databases up to date.