Refactored so that the frontend loads api message codes dynamically
Instead of trying to keep the api message codes in sync between the
front and back ends, instead, the front-end loads a mapping of names to
api message codes from the backend when it first connects.  This only
requires the use of a single, static message code for the initial api
request/response pair, after which the backend is the single source of
truth for api codes.