Fix a card importer crash when version from mtgjson doesn't have a date
The JSON unmarshalling code for the mtgjson version structure assumed
that the version field would always have the date offset in addition to
the semver version.  This isn't true for the first release of a new
version, where the version field just has the semver version without a
date offset.  This caused a crash in the JSON unmarshaller.  Since this
is a valid format for the version field, this fixes the version JSON
unmarshaller to accept this format as well.