Rewrote card hashing implementation, fixed several bugs
The existing card hashing implementation was based on handwritten code
that manually added every struct field to the hash.  This is error prone
when adding new fields, it would be easy to forget to add a new field to
the hash, leading to the hash becoming unstable.  Rewrote the hash
function to be generic using golang reflection, which in addition to
handling new fields automatically, removed a lot of repetitive code.

Also added a unit test for hash stability with the new hash function,
which exposed several bugs in the stability of the existing
implementation, that this also fixes.  The first issue is that several
of the sorting operations (to make sure the data was in a consistent
order before hashing) were sorting on fields that could contain
duplicates.  This had the effect of making the sort order within the
sub-groups of items with the same key unstable.  Fixed by updating all
of the sorting functions to sort hierarchically within groups that
contained duplicate keys, using other fields to disambiguate.

Another issue uncovered is that some double-sided tokens that have the
same name on both sides share a UUID.  This broke the assumption that
all card faces have a unique UUID.  Will follow-up with the MTGJSON
project to see if that's intended or not.